+++ boiduongLeQuyDon/boiduongLeQuyDon/GUI/QuanLyPhuDao.cs
@@ -56,7 +56,8 @@ namespace boiduongLeQuyDon.GUI
 
         private void simpleButton3_Click(object sender, EventArgs e)
         {
-            queries.deletePhuDao(Convert.ToInt32(lblID.Text));
+            if (MessageBox.Show("Bạn có thật sự muốn xóa?", "Có", MessageBoxButtons.YesNo) == System.Windows.Forms.DialogResult.Yes)
+                queries.deletePhuDao(Convert.ToInt32(lblID.Text));
             load();
         }
